Added all components Added Build functionality Added build compatibility checking functionality

+++ src/main/java/com/pcplanet/pcplanetbackend/component/gpu/output_interface/GPUOutputInterface.java
@@ -13,8 +13,9 @@ import lombok.NoArgsConstructor;
 public class GPUOutputInterface {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Short id;
     private String interfaceName;
+
     public GPUOutputInterface(String interfaceName) {
         this.interfaceName = interfaceName;
     }
